KeepFit
This is a fitness app What does this application do -it shows a number
of different exercises which are added in a list fiew to dislpay the
information -it displays a map control allowing you to see gyms around
galway. -it displays a few choices of food and what the calories for
each are and you can also calculate them. -it shows a couple of
chalenges that you can try. -and finally it displays the streatches
nesessary after the workout